% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fst_probs.R
\name{read_fst_probs}
\alias{read_fst_probs}
\title{Read probs matrix from fst}
\usage{
read_fst_probs(prefix)
}
\arguments{
\item{prefix}{prefix passed to write_fst_probs()}
}
\description{
Given some fst files written by write_fst_probs, read them back into a 3D probs array. Took ~5s on my 2-core early 2015 Macbook Pro with SSD for probs with dimension 185 * 8 * 69005
}
