% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fst_probs.R
\name{write_fst_probs}
\alias{write_fst_probs}
\title{Write probs matrix to fst}
\usage{
write_fst_probs(probs, prefix, compress = 50)
}
\arguments{
\item{probs}{3D matrix of probs in DOQTL format}

\item{prefix}{prefix for fst files}

\item{compress}{compression level for fst files (passed to fst function)}
}
\description{
This function writes a probs matrix to a series of fst files. Took ~10s on my 2-core early 2015 Macbook Pro with SSD for probs with dimension 185 * 8 * 69005
}
